% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_api.R
\name{ns_api_fetch}
\alias{ns_api_fetch}
\title{Generate a single Nightscout API request, and perform request to get response}
\usage{
ns_api_fetch(
  hostname = NULL,
  pathname = "api/v1",
  filename = c("entries", "treatments", "devicestatus", "profile", "status"),
  filetype = c("json", "csv"),
  count = 10,
  start_date = NULL,
  end_date = NULL,
  api_secret = NULL,
  api_hash = NULL,
  path = NULL,
  perform = TRUE,
  ...
)
}
\arguments{
\item{hostname}{Base URL for request.}

\item{pathname}{Default api/v1}

\item{filename}{Data collection to retrieve. CGM data = \code{entries}, Care Portal Treatments = \code{treatments}, Treatment profiles = \code{profile}, Server status and settings = \code{status}}

\item{filetype}{Data to be return as \code{json} or \code{csv}}

\item{count}{Number of values to return. Default is 10, set to \code{Inf} to ensure all are returned when date range is specifed.}

\item{start_date}{First day of date range to return. As date field or character in form YYYY-MM-DD. \strong{Note} The extract will always start at midnight UTC / GMT as that is how the data is saved within Nightscout.}

\item{end_date}{Last day of date range to return. As date field or character in form YYYY-MM-DD. \strong{Note} The extract will always end at 23:59:59 UTC / GMT as that is how the data is saved within Nightscout.}

\item{api_secret}{API secret / user password.}

\item{api_hash}{Hashed version of API secret / user password. If supplied api_secret is ignored.}

\item{path}{Passed to httr2::req_perform. Path to save body of the response (data retrieved). This is useful for large responses since it avoids storing the response in memory. Optional, if not supplied data will be directly returned into R. If path ends with .Rds will save as an R data file with full httr2 repsonse information.}

\item{perform}{Whether to perform to request \code{TRUE} or just return the request itself \code{FALSE}}
}
\value{
Response from httr2::req_perform

If the HTTP request succeeds, and the status code is ok (e.g. 200), an HTTP response.

If the HTTP request succeeds, but the status code is an error (e.g a 404), an error with class c("httr2_http_404", "httr2_http"). By default, all 400 and 500 status codes will be treated as an error, but you can customise this with req_error().
If the HTTP request fails (e.g. the connection is dropped or the server doesn't exist), an error with class "httr2_failure".
}
\description{
Generate a single Nightscout API request, and perform request to get response
}
\examples{
\dontrun{
ns_api_fetch(url = "http://localhost.1337", filename = "entries", count = 100)
}


}
