% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_basal_temporary}
\alias{ns_basal_temporary}
\title{Create data.frame of temporary basal insulin rates by day and time}
\usage{
ns_basal_temporary(treatments)
}
\arguments{
\item{treatments}{\code{data.table} of cleaned Nightscout Care Portal Treatment information.}
}
\value{
An R \code{data.table} with columns by (day), start_epoch, end_epoch, and rate (units per hour).
}
\description{
Create data.frame of temporary basal insulin rates by day and time
}
