% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_basal_total}
\alias{ns_basal_total}
\title{Create data.frame of combined scheduled and temporary basal insulin units delivered by day.}
\usage{
ns_basal_total(treatments)
}
\arguments{
\item{treatments}{\code{data.table} of cleaned Nightscout Care Portal Treatment information.}
}
\value{
An R \code{data.table} with columns by (day), start_epoch, end_epoch, and units of insulin scheduled and delivered.
}
\description{
Create data.frame of combined scheduled and temporary basal insulin units delivered by day.
}
