% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_bolus_total}
\alias{ns_bolus_total}
\title{Extract insulin bolus information from Nightscout treatment data}
\usage{
ns_bolus_total(treatments)
}
\arguments{
\item{treatments}{\code{data.table} of cleaned Nightscout Care Portal Treatment information.}
}
\value{
An R \code{data.table} with columns by (day), time, bolus type (SMB or manual) and units of insulin delivered.
}
\description{
Extract insulin bolus information from Nightscout treatment data
}
