% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_api.R
\name{ns_clean_treatments}
\alias{ns_clean_treatments}
\title{Clean treatment data. Mostly creating local date variables.}
\usage{
ns_clean_treatments(treatments, tz = NULL)
}
\arguments{
\item{treatments}{\code{data.table} of raw Nightscout Care Portal Treatment information}

\item{tz}{If not provided will take the most common timezone in treatments.}
}
\value{
The data.table is modified by reference (and returned invisibly).
}
\description{
Clean treatment data. Mostly creating local date variables.
}
