% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_date_range}
\alias{ns_date_range}
\title{Create dataset from date range}
\usage{
ns_date_range(x, tz = NULL)
}
\arguments{
\item{x}{Vector of at least two dates or datetimes, to identify the minimum and maximum dates to be included.}

\item{tz}{If x is a date object, a timezone also needs to be provided.}
}
\value{
A data.table with columns by (day) and unix epoch for start of day (in local timezone).
}
\description{
Create dataset from date range
}
