% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_extract_profiles}
\alias{ns_extract_profiles}
\title{Extract dates and times of profile switches, and scheduled basals, from Nightscout treatment information}
\usage{
ns_extract_profiles(treatments)
}
\arguments{
\item{treatments}{\code{data.table} of cleaned Nightscout Care Portal Treatment information.}
}
\value{
A \code{data.table} with dates and times of profile switches
}
\description{
Extract dates and times of profile switches, and scheduled basals, from Nightscout treatment information
}
