% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_insulin}
\alias{ns_insulin}
\title{Create data.frame of basal and bolus insulin scheduled and delivered by day}
\usage{
ns_insulin(treatments)
}
\arguments{
\item{treatments}{\code{data.table} of cleaned Nightscout Care Portal Treatment information.}
}
\value{
An R \code{data.table} with columns by (day), basal and bolus insulin delivered to match Nightscout dashboard. Calculations aren't exactly the same, not sure why?
}
\description{
Create data.frame of basal and bolus insulin scheduled and delivered by day
}
