% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_api.R
\name{ns_read_file}
\alias{ns_read_file}
\title{Read in json or Rds data extract from Nightscout previously and saved on a disk}
\usage{
ns_read_file(file)
}
\arguments{
\item{file}{A json or Rds file previous saved by function \code{ns_retrieve}, or other httr2::req_perform operation.}
}
\value{
Data table
}
\description{
Read in json or Rds data extract from Nightscout previously and saved on a disk
}
