% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_requests.R
\name{run_requests}
\alias{run_requests}
\title{Wrapper to create and run Nightscout API requests and return data to global environment}
\usage{
run_requests(type = "json", ...)
}
\arguments{
\item{type}{Type of data file to output. 'json' as extracted from Nightscout, or 'data' to convert to r \code{data.table} using \code{data.table::rbindlist}.}

\item{...}{Additional arguments to pass to \code{make_requests}. Including base \code{url} and \code{api_secret}.}
}
\description{
Wrapper to create and run Nightscout API requests and return data to global environment
}
