% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_requests.R
\name{save_requests}
\alias{save_requests}
\title{Create and run Nightscout API requests and save returned data to a specified external folder}
\usage{
save_requests(path = "", suffix = "", type = "json", ...)
}
\arguments{
\item{path}{A folder location where retrieved data is to be saved.}

\item{suffix}{By default files are saved according to their schema (e.g. entries.Rds), a suffix may be applied to the filename.}

\item{type}{Type of data file to output. 'json' as extracted from Nightscout, or 'data' to convert to r \code{data.table} using \code{data.table::rbindlist}.}

\item{...}{Additional arguments to pass to \code{make_requests}. Including base \code{url} and \code{api_secret}.}
}
\description{
Create and run Nightscout API requests and save returned data to a specified external folder
}
