% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightscout_insulin.R
\name{ns_date_range}
\alias{ns_date_range}
\title{Create dataset from date range}
\usage{
ns_date_range(
  treatments,
  date_var = "dttm_local",
  date_range = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{treatments}{\code{data.table} of cleaned Nightscout Care Portal Treatment information.}

\item{date_var}{Name of column containing date variable.}

\item{date_range}{If date minimum and maximum are not provided, will obtain these from the treatments}

\item{tz}{Expected timezone of users data. Defaults to system timezone.}
}
\value{
A data.table with columns by (day) and unix epoch for start of day.
}
\description{
Create dataset from date range
}
